Change of color scheme
It was pointed out to me that the old one is a little hard to read due
to the contrast (thanks to citrons for pointing this out to me), so
i've changed it to this one.  This is significantly more blue and the
 text is now bright instead of dark, and now it kinda looks like an
old-school TUI, which is kinda fun.

I have been thinking of using an actually professional color scheme
instead of just web colors.  I was thinking of Solarized, but i have
several problems with it, and honestly i'm thinking of just making my
own color scheme.  If i do, i'll probably change my page to use it.