Farewell, GitHub
Yet another once-great creature fallen to the hand of capitalism in general and Microsoft specifically.  This is now the second time that i've made an account on something that then got purchased by Microsoft, who then fucked with the account system in ways i found unnacceptable to the point of choosing to abandon it.